src/requirements.txt
numpy
opencv-python
matplotlib
qiskit
qiskit-aer
qiskit-ibm-runtime
pycryptodome
tk
